--- src/main/java/POT/DuoBloom/community/dto/CommunityCommentRequestDto.java
@@ -1,13 +1,0 @@
-package POT.DuoBloom.community.dto;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class CommunityCommentRequestDto {
-    private String content;
-}
