+++ src/main/java/POT/DuoBloom/community/dto/request/CommunityCommentRequestDto.java
@@ -1,4 +1,5 @@
-package POT.DuoBloom.community.dto;
+package POT.DuoBloom.community.dto.request;
+
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,7 +8,6 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class CommunityImageResponseDto {
-    private Long imageId;
-    private String imageUrl;
+public class CommunityCommentRequestDto {
+    private String content;
 }
