+++ src/main/java/POT/DuoBloom/community/dto/request/CommunityRequestDto.java
@@ -1,4 +1,4 @@
-package POT.DuoBloom.community.dto;
+package POT.DuoBloom.community.dto.request;
 
 import POT.DuoBloom.community.entity.Type;
 import lombok.AllArgsConstructor;
@@ -8,10 +8,8 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class CommunityResponseDto {
-    private Long communityId;
+public class CommunityRequestDto {
     private String content;
     private Type type;
-    private String nickname;
-    private String profilePictureUrl;
 }
+
