+++ src/main/java/POT/DuoBloom/community/dto/response/CommunityImageResponseDto.java
@@ -1,0 +1,13 @@
+package POT.DuoBloom.community.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class CommunityImageResponseDto {
+    private Long imageId;
+    private String imageUrl;
+}
