+++ src/main/java/POT/DuoBloom/community/dto/response/CommunityResponseDto.java
@@ -1,4 +1,4 @@
-package POT.DuoBloom.community.dto;
+package POT.DuoBloom.community.dto.response;
 
 import POT.DuoBloom.community.entity.Type;
 import lombok.AllArgsConstructor;
@@ -8,8 +8,10 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class CommunityRequestDto {
+public class CommunityResponseDto {
+    private Long communityId;
     private String content;
     private Type type;
+    private String nickname;
+    private String profilePictureUrl;
 }
-
